backend/requirements.txt
tensorflow>=2.12.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
Pillow>=10.0.0
